chore: addressed comments on PR
