+++ apps/academy-wep/app/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import WhyChoose from '@/components/sections/why-choose'
+import { WhyChoose } from '@/components/sections/why-choose'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { motion } from 'framer-motion'
